+++ R/imputation.R
@@ -6,11 +6,11 @@
 ##' objects instance using a variety of methods (see below).
 ##'
 ##' Users should proceed with care when imputing data and take
-##' precautions to assure that the imputation produce valid results,
+##' precautions to assure that the imputation produces valid results,
 ##' in particular with naive imputations such as replacing missing
 ##' values with 0.
 ##'
-##' @details
+##' @section Types of missing values:
 ##'
 ##' There are two types of mechanisms resulting in missing values in
 ##' LC/MSMS experiments.
@@ -18,22 +18,50 @@
 ##' - Missing values resulting from absence of detection of a feature,
 ##'   despite ions being present at detectable concentrations. For
 ##'   example in the case of ion suppression or as a result from the
-##'   stochastic, data-dependent nature of the MS acquisition
+##'   stochastic, data-dependent nature of the DDA MS acquisition
 ##'   method. These missing value are expected to be randomly
-##'   distributed in the data and are defined as missing at random
-##'   (MAR) or missing completely at random (MCAR).
+##'   distributed in the data and are defined, in statistical terms,
+##'   as missing at random (MAR) or missing completely at random (MCAR).
 ##'
 ##' - Biologically relevant missing values resulting from the absence
-##'   of the low abundance of ions (below the limit of detection of
+##'   or the low abundance of ions (i.e. below the limit of detection of
 ##'   the instrument). These missing values are not expected to be
 ##'   randomly distributed in the data and are defined as missing not
 ##'   at random (MNAR).
 ##'
 ##' MNAR features should ideally be imputed with a left-censor method,
-##' such as `QRILC` below. Conversely, it is recommended to use host
+##' such as `QRILC` below. Conversely, it is recommended to use hot
 ##' deck methods such nearest neighbours, Bayesian missing value
 ##' imputation or maximum likelihood methods when values are missing
 ##' at random.
+##'
+##' @section Imputing by rows or columns:
+##'
+##' We assume that the input matrix `x` contains features along the
+##' rows and samples along the columns, as is generally the case in
+##' omics data analysis. When performing imputation, the missing
+##' values are taken as a feature-specific property: feature *x* is
+##' missing because it is absent (in a sample or group), or because it
+##' was missed during acquisition (not selected during data dependent
+##' acquisition) or data processing (not identified or with an
+##' identification score below a chosen false discovery threshold). As
+##' such, imputation is by default performed at the *feature
+##' level*. In some cases, such as imputation by zero or a global
+##' minimum value, it doesn't matter. In other cases, it does matter
+##' very much, such as for example when using the minimum value
+##' computed for each margin (i.e. row or column) as in the *MinDet*
+##' method (see below) - do we want to use the minimum of the sample
+##' or of that features? KNN is another such example: do we consider
+##' the most similar features to impute a feature with missing values,
+##' or the most similar samples to impute all missing in a sample.
+##'
+##' The `margin` argument can be used to change the imputation margin
+##' from features/rows (`margin = 1`) to samples/columns (`margin = 2`).
+##' Different imputations will have different default values, and
+##' changing this parameter can have a major impact on imputation results
+##' and downstream results.
+##'
+##' @section Imputation methods:
 ##'
 ##' Currently, the following imputation methods are available.
 ##'
@@ -50,7 +78,7 @@
 ##' - *RF*: Random Forest imputation, as implemented in the
 ##'   `missForest::missForest` function. See [missForest::missForest()]] for
 ##'   details and additional parameters.
-##'   
+##'
 ##' - *knn*: Nearest neighbour averaging, as implemented in the
 ##'   `impute::impute.knn` function. See [impute::impute.knn()]] for
 ##'   details and additional parameters.
@@ -58,10 +86,8 @@
 ##' - *QRILC*: A missing data imputation method that performs the
 ##'   imputation of left-censored missing data using random draws from
 ##'   a truncated distribution with parameters estimated using
-##'   quantile regression. Implemented in the
-##'   `imputeLCMD::impute.QRILC`
-##'   function. [imputeLCMD::impute.QRILC()] for details and
-##'   additional parameters.
+##'   quantile regression. The `impute_QRILC()` function calls
+##'   [imputeLCMD::impute.QRILC()] from the `imputeLCMD` package.
 ##'
 ##' - *MinDet*: Performs the imputation of left-censored missing data
 ##'   using a deterministic minimal value approach. Considering a
@@ -69,10 +95,8 @@
 ##'   sample, the missing entries are replaced with a minimal value
 ##'   observed in that sample. The minimal value observed is estimated
 ##'   as being the q-th quantile (default `q = 0.01`) of the observed
-##'   values in that sample. Implemented in the
-##'   `imputeLCMD::impute.MinDet` function. See
-##'   [imputeLCMD::impute.MinDet()] for details and additional
-##'   parameters.
+##'   values in that sample. The implemented in based on the
+##'   `imputeLCMD::impute.MinDet()` function.
 ##'
 ##' - *MinProb*: Performs the imputation of left-censored missing data
 ##'   by random draws from a Gaussian distribution centred to a
@@ -82,12 +106,11 @@
 ##'   sample. The minimal value observed is estimated as being the
 ##'   q-th quantile (default `q = 0.01`) of the observed values in
 ##'   that sample. The standard deviation is estimated as the median
-##'   of the feature standard deviations. Note that when estimating
-##'   the standard deviation of the Gaussian distribution, only the
-##'   peptides/proteins which present more than 50\% recorded values
-##'   are considered. Implemented in the `imputeLCMD::impute.MinProb`
-##'   function. See [imputeLCMD::impute.MinProb()] for details and
-##'   additional parameters.
+##'   of the feature (or sample) standard deviations. Note that when
+##'   estimating the standard deviation of the Gaussian distribution,
+##'   only the peptides/proteins which present more than 50\% recorded
+##'   values are considered. The `impute_MinProb()` function calls
+##'   [imputeLCMD::impute.MinProb()] from the `imputeLCMD` package.
 ##'
 ##' - *min*: Replaces the missing values with the smallest non-missing
 ##'   value in the data.
@@ -119,7 +142,8 @@
 ##'   method.
 ##'
 ##' The `imputeMethods()` function returns a vector with valid
-##' imputation method arguments.
+##' imputation method names. Use `getImputeMargin()` to get the
+##' default margin for each imputation function.
 ##'
 ##' @references
 ##'
@@ -143,7 +167,7 @@
 ##'
 ##' @rdname imputation
 ##'
-##' @aliases imputeMethods impute_neighbour_average impute_knn impute_mle impute_bpca impute_mixed impute_min impute_zero impute_with impute_matrix
+##' @aliases imputeMethods impute_neighbour_average impute_knn impute_mle impute_bpca impute_mixed impute_min impute_zero impute_with impute_matrix impute_MinDet impute_MinProb impute_QRILC
 ##'
 ##' @useDynLib MsCoreUtils, .registration = TRUE
 ##'
@@ -191,8 +215,15 @@
 ##' @param ... Additional parameters passed to the inner imputation
 ##'     function.
 ##'
+##' @param margin `integer(1)` defining the margin along which to
+##'     apply imputation, with `1L` for rows and `2L` for columns. The
+##'     default value will depend on the imputation method. Use
+##'     `getImputeMargin(fun)` to get the default margin of imputation
+##'     function `fun`. If the function doesn't take a margin
+##'     argument, `NA` is returned.
+##'
 ##' @return A matrix of same class as `x` with dimensions `dim(x)`.
-##' 
+##'
 ##' @export
 impute_matrix <- function(x,
                           method,
@@ -205,7 +236,7 @@ impute_matrix <- function(x,
         xIsHDF5 <- TRUE
         p <- HDF5Array::path(x) ## stored for later writing to disk
         ## Watch out this can lead to memory burst when x is large.
-        x <- as.matrix(x) 
+        x <- as.matrix(x)
     }
     ## User-provided imputation function
     if (!missing(FUN) && is.function(FUN)) {
@@ -224,7 +255,7 @@ impute_matrix <- function(x,
                         choices = imputeMethods(),
                         several.ok = FALSE)
     res <- x
-    if (method %in% c("CRILC", "MinDet", "MinProb"))
+    if (method %in% c("QRILC", "MinProb"))
         requireNamespace("imputeLCMD")
     if (method == "knn") {
         res <- impute_knn(x, ...)
@@ -232,14 +263,16 @@ impute_matrix <- function(x,
         res <- impute_neighbour_average(x, ...)
     } else if (method == "MLE") {
         res <- impute_mle(x, ...)
+    } else if (method == "MLE2") {
+        res <- impute_mle2(x, ...)
     } else if (method == "bpca"){
         res <- impute_bpca(x, ...)
+    } else if (method == "MinDet") {
+        res <- impute_MinDet(x, ...)
+    } else if (method == "MinProb") {
+        res <- impute_MinProb(x, ...)
     } else if (method == "QRILC") {
         res <- imputeLCMD::impute.QRILC(x, ...)[[1]]
-    } else if (method == "MinDet") {
-        res <- imputeLCMD::impute.MinDet(x, ...)
-    } else if (method == "MinProb") {
-        res <- imputeLCMD::impute.MinProb(x, ...)
     } else if (method == "min") {
         res <- impute_min(x)
     } else if (method == "mixed") {
@@ -263,7 +296,7 @@ impute_matrix <- function(x,
 ##' @export
 ##' @rdname imputation
 imputeMethods <- function()
-    c("bpca","knn", "QRILC", "MLE",
+    c("bpca","knn", "QRILC", "MLE", "MLE2",
       "MinDet", "MinProb", "min", "zero",
       "mixed", "nbavg", "with", "RF", "none")
 
@@ -272,53 +305,95 @@ imputeMethods <- function()
 ##' @param k `numeric(1)` providing the imputation value used for the
 ##'     first and last samples if they contain an `NA`. The default is
 ##'     to use the smallest value in the data.
-impute_neighbour_average <- function(x, k = min(x, na.rm = TRUE)) {
+impute_neighbour_average <- function(x, k = min(x, na.rm = TRUE), margin = 1L) {
     message("Assuming values are ordered.")
-    .Call("C_impNeighbourAvg", x, k)
+    margin <- .checkMargin(margin)
+    if (margin == 2L) x <- t(x)
+    res <- .Call("C_impNeighbourAvg", x, k)
+    if (margin == 2L) return(t(res))
+    else return(res)
 }
 
 ##' @export
 ##' @rdname imputation
-impute_knn <- function(x, ...) {
+impute_knn <- function(x, margin = 1L, ...) {
     requireNamespace("impute")
+    margin <- .checkMargin(margin)
+    if (margin == 2L)
+        x <- t(x)
     imp_res <- impute::impute.knn(x, ...)
     if (!is.null(imp_res$rng.state)) {
         assign(".Random.seed", imp_res$rng.state, envir = .GlobalEnv)
     } else {
         rm(".Random.seed", envir = .GlobalEnv)
     }
-    imp_res$data
+    if (margin == 2L)
+        return(t(imp_res$data))
+    else imp_res$data
 }
 
 ##' @export
 ##' @rdname imputation
-impute_mle <- function(x, ...) {
+impute_mle <- function(x, margin = 2L, ...) {
     requireNamespace("norm")
+    margin <- .checkMargin(margin)
+    if (margin == 2L)
+        x <- t(x)
     s <- norm::prelim.norm(x)  ## preliminary manipulations
     th <- norm::em.norm(s, ...) ## find the MLE
     seed <- sample(.Machine$integer.max, 1)
     norm::rngseed(seed) ## set random number generator seed
-    norm::imp.norm(s, th, x)  ## impute missing data under the MLE
+    res <- norm::imp.norm(s, th, x)  ## impute missing data under the MLE
+    if (margin == 2L)
+        res <- t(res)
+    res
 }
 
 ##' @export
 ##' @rdname imputation
-impute_bpca <- function(x, ...) {
+impute_mle2 <- function(x, margin = 2L, ...) {
+    requireNamespace("norm2")
+    margin <- .checkMargin(margin)
+    dn <- dimnames(x)
+    if (margin == 2L)
+        x <- t(x)
+    res <- norm2::emNorm(obj = x, ...)$y.mean.imp
+    if (margin == 2L)
+        res <- t(res)
+    dimnames(res) <- dn
+    res
+}
+
+##' @export
+##' @rdname imputation
+impute_bpca <- function(x, margin = 1L, ...) {
     requireNamespace("pcaMethods")
+    margin <- .checkMargin(margin)
+    if (margin == 2L)
+        x <- t(x)
     nSamples <- dim(x)[2]
     .resultBPCA <- pcaMethods::pca(x,
                                    method = "bpca",
                                    nPcs = (nSamples - 1),
                                    verbose = FALSE,
                                    ...)
-    pcaMethods::completeObs(.resultBPCA)
+    res <- pcaMethods::completeObs(.resultBPCA)
+    if (margin == 2L)
+        res <- t(res)
+    res
 }
 
 ##' @export
 ##' @rdname imputation
-impute_RF <- function(x, ...) {
+impute_RF <- function(x, margin = 2L, ...) {
     requireNamespace("missForest")
-    t(missForest::missForest(t(x), ...)$ximp)
+    margin <- .checkMargin(margin)
+    if (margin == 1L)
+        x <- t(x)
+    res <- missForest::missForest(x, ...)$ximp
+    if (margin == 1L)
+        res <- t(res)
+    res
 }
 
 ##' @param randna `logical` of length equal to `nrow(object)` defining
@@ -335,7 +410,7 @@ impute_RF <- function(x, ...) {
 ##' @export
 ##'
 ##' @rdname imputation
-impute_mixed <- function(x, randna, mar, mnar, ...) {
+impute_mixed <- function(x, randna, mar, mnar, margin = 1L, ...) {
     if (missing(randna))
         stop("Mixed imputation requires 'randna' argument. See ?impute_mixed.",
              call. = FALSE)
@@ -346,11 +421,16 @@ impute_mixed <- function(x, randna, mar, mnar, ...) {
     if (missing(mnar))
         stop("Mixed imputation requires 'mnar' argument. See ?impute_mixed.",
              call. = FALSE)
+    suppressMessages(margin <- .checkMargin(margin))
+    if (margin == 2L) x <- t(x)
     if (length(randna) != nrow(x))
         stop("Number of proteins and length of randna must be equal.",
              call. = FALSE)
-    x[randna, ] <- impute_matrix(x[randna, ], mar, ...)
-    x[!randna, ] <- impute_matrix(x[!randna, ], mnar, ...)
+    x[randna, ] <- impute_matrix(x[randna, ], mar,
+                                 margin = margin, ...)
+    x[!randna, ] <- impute_matrix(x[!randna, ], mnar,
+                                  margin = margin, ...)
+    if (margin == 2L) x <- t(x)
     x
 }
 
@@ -366,11 +446,55 @@ impute_min <- function(x) {
 ##' @export
 ##'
 ##' @rdname imputation
+##'
+##' @importFrom stats quantile
+##'
+##' @param q `numeric(1)` indicating the quantile to be used to
+##'     estimate the minimum in `MinDet` and `MinProb`. Default is
+##'     0.01.
+impute_MinDet <- function(x, q = 0.01, margin = 2L) {
+    margin <- .checkMargin(margin)
+    n <- dim(x)[margin]
+    impVals <- apply(x, margin, quantile, prob = q, na.rm = TRUE)
+    for (i in seq_len(n))
+        x[is.na(x[, i]), i] = impVals[i]
+    x
+}
+
+##' @export
+##'
+##' @rdname imputation
+##'
+##' @param sigma `numeric(1)` controling the standard deviation of the
+##'     MNAR distribution in `MinProb` and `QRILC`. Default is 1.
+impute_MinProb <- function (x, q = 0.01, sigma = 1, margin = 2L) {
+    margin <- .checkMargin(margin)
+    ## transpose if we want to non-default version
+    if (margin == 1L) x <- t(x)
+    x <- imputeLCMD::impute.MinProb(x, q = q, tune.sigma = sigma)
+    if (margin == 1L) x <- t(x)
+    x
+}
+
+##' @export
+##'
+##' @rdname imputation
+impute_QRILC <- function(x, sigma = 1, margin = 2L) {
+   margin <- .checkMargin(margin)
+    ## transpose if we want to non-default version
+    if (margin == 1L) x <- t(x)
+    x <- imputeLCMD::impute.QRILC(x, tune.sigma = sigma)[[1]]
+    if (margin == 1L) x <- t(x)
+    x
+}
+
+##' @export
+##'
+##' @rdname imputation
 impute_zero <- function(x) {
     x[is.na(x)] <- 0
     x
 }
-
 
 ##' @export
 ##'
@@ -403,8 +527,46 @@ impute_with <- function(x, val) {
 ##' }
 ##'
 ##' impute_matrix(m, FUN = random_imp)
-impute_fun <- function(x, FUN, ...) {
+impute_fun <- function(x, FUN, margin = 1L, ...) {
+    margin <- .checkMargin(margin)
+    if (margin == 2L) x <- t(x)
     res <- do.call(FUN, list(x, ...))
+    if (margin == 2L) res <- t(res)
     stopifnot(identical(dim(x), dim(res)))
     res
+}
+
+.checkMargin <- function(margin) {
+    margin <- as.integer(margin)
+    if (!margin %in% c(1L, 2L))
+        stop("'margin' must be 1L or 2L")
+    message("Imputing along margin ", margin,
+            ifelse(margin == 1L, " (features/rows)",
+                   " (samples/columns)"), ".")
+    margin
+}
+
+##' @export
+##'
+##' @rdname imputation
+##'
+##' @param fun The imputation function to get the default margin from.
+##'
+##' @examples
+##'
+##' ## get the default margin
+##' getImputeMargin(impute_knn) ## default imputes along features
+##'
+##' getImputeMargin(impute_mle) ## default imputes along samples
+##'
+##' getImputeMargin(impute_zero) ## NA: no margin here
+##'
+##' ## default margin for all MsCoreUtils::impute_* functions
+##' sapply(ls("package:MsCoreUtils", pattern = "impute_"), getImputeMargin)
+getImputeMargin <- function(fun) {
+    args <- formals(fun)
+    i <- grep("margin", names(args))
+    if (length(i)) ans <- args[[i]]
+    else ans <- NA
+    ans
 }